SmartSolarUnit/solar_python_engine/requirements.txt
pandas
numpy
joblib
firebase-admin
python-dotenv
xgboost
flask
shap
prophet
statsmodels